+++ DRInformationSystem/BusinessLogicLayer/Services/ResponsesService/ResponsesService.cs
@@ -1,4 +1,5 @@
-﻿using BusinessLogicLayer.Models;
+﻿using BusinessLogicLayer.Exceptions;
+using BusinessLogicLayer.Models;
 using DataAccessLayer.Models;
 using DataAccessLayer.Repositories;
 
@@ -7,10 +8,16 @@ namespace BusinessLogicLayer.Services;
 public class ResponsesService : IResponsesService
 {
 	private readonly IResponsesRepository _responsesRepository;
+	private readonly IOrdersRepository _ordersRepository;
+	private readonly IConsumersRepository _consumersRepository;
 
-	public ResponsesService(IResponsesRepository responsesRepository)
+	public ResponsesService(IResponsesRepository responsesRepository,
+		IOrdersRepository ordersRepository,
+		IConsumersRepository consumersRepository)
 	{
 		_responsesRepository = responsesRepository;
+		_ordersRepository = ordersRepository;
+		_consumersRepository = consumersRepository;
 	}
 
 	public async Task<List<ResponseModel>> GetResponsesForOrderAsync(long orderId)
@@ -35,6 +42,14 @@ public class ResponsesService : IResponsesService
 
 	public async Task CreateResponseAsync(CreateResponseData createData)
 	{
+		var order = await _ordersRepository.GetByIdAsync(createData.OrderId);
+		if (order == null)
+			throw new NotExistedOrderException(createData.OrderId);
+
+		var consumer = await _consumersRepository.GetByIdAsync(createData.ConsumerId);
+		if (consumer == null)
+			throw new NotExistedConsumerException(createData.ConsumerId);
+
 		var responseModel = new ResponseModel
 		{
 			ConsumerId = createData.ConsumerId,
@@ -48,6 +63,8 @@ public class ResponsesService : IResponsesService
 	public async Task DeleteResponseAsync(long responseId)
 	{
 		var response = await _responsesRepository.GetByIdAsync(responseId);
+		if (response == null)
+			throw new NotExistedResponseException(responseId);
 
 		await _responsesRepository.DeleteAsync(response);
 	}
